The MazeGame can now be given a custom width, height, and seed, and will make the maze accordingly. The new custom UI allows that to be supplied. If the seed is "0", it'll be random. Every other one is set. If a value that's not a number is given to any of these things, it'll spew exceptions, but it won't kill the program or affect the execution which is alright, but should be cleaned up.

+++ src/Maze.java
@@ -57,7 +57,7 @@ public class Maze {
 	 * A specific seed for the random generation (the same seed will give the
 	 * same results).
 	 */
-	public void generateMapDepthStyle(int seed) {
+	public void generateMapDepthStyle(long seed) {
 		generateMapDepthStyleProcess(new Random(seed));
 	}
